api/schemas: Merge into "skylines.schemas" module
