+++ ember/app/templates/components/upload-result-form.hbs
@@ -129,7 +129,7 @@
         {{#each result.airspaces as |airspace|}}
           <tr>
             <td class="name">{{airspace.name}}</td>
-            <td class="class">{{airspace.type}}</td>
+            <td class="class">{{airspace.class}}</td>
             <td class="base">{{airspace.base}}</td>
             <td class="top">{{airspace.top}}</td>
             <td>
