--- skylines/api/schemas/__init__.py
@@ -1,7 +1,0 @@
-from .airspace import AirspaceSchema
-from .wave import WaveSchema
-
-__all__ = [
-    AirspaceSchema,
-    WaveSchema,
-]
