+++ tests/api/schemas/airspace_test.py
@@ -4,13 +4,13 @@ from datetime import datetime
 from geoalchemy2.shape import to_shape
 from shapely.geometry import Polygon
 
-from skylines.api import schemas
+from skylines.schemas import AirspaceSchema
 
 
 def test_airspace_schema(test_airspace):
     """:type test_airspace: skylines.model.Airspace"""
 
-    data, errors = schemas.AirspaceSchema().dump(test_airspace)
+    data, errors = AirspaceSchema().dump(test_airspace)
     assert not errors
 
     assert isinstance(data, OrderedDict)
@@ -18,10 +18,10 @@ def test_airspace_schema(test_airspace):
         'id',
         'name',
         'class',
-        'top',
         'base',
+        'top',
         'shape',
-        'country',
+        'countryCode',
         'created_at',
         'modified_at',
     ]
@@ -31,7 +31,7 @@ def test_airspace_schema(test_airspace):
     assert data['class'] == test_airspace.airspace_class
     assert data['top'] == test_airspace.top
     assert data['base'] == test_airspace.base
-    assert data['country'] == test_airspace.country_code
+    assert data['countryCode'] == test_airspace.country_code
 
     created_at = datetime.strptime(data['created_at'], '%Y-%m-%dT%H:%M:%S.%f+00:00')
     assert isinstance(created_at, datetime)
@@ -51,5 +51,5 @@ def test_airspace_schema(test_airspace):
     c2 = shape[1]
     test_c2 = test_shape.exterior.coords[1]
 
-    assert c2['longitude'] == test_c2[0]
-    assert c2['latitude'] == test_c2[1]
+    assert c2[0] == test_c2[0]
+    assert c2[1] == test_c2[1]
